pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["process", "workflows", "qdrant_storage"]
exclude = ["data*", "logs*", "tests*"]
namespaces = false

[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG System"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.15",
    "qdrant-client>=1.6.9",
    "pypdf>=3.15.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.1",
    "sentence-transformers>=2.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "mem0ai>=0.1.116",
    "chainlit>=2.7.1.1",
    "tokenizers>=0.21.4",
    "streamlit>=1.49.1",
    "pymupdf>=1.26.4",
    "langchain-experimental>=0.3.4",
    "unstructured>=0.18.14",
    "python-magic>=0.4.27",
    "python-pptx>=1.0.2",
    "pytesseract>=0.3.13",
    "pyppeteer>=0.0.25",
    "websockets<12",
    "grandalf>=0.8",
    "logfire>=4.10.0",
    "deepeval>=2.6.7",
    "langchain-anthropic>=0.3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "ipython>=8.0.0",
    "pylint>=2.17.0"
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
